Section 4 Unify the interface with tabs You have two parallel interfaces to view and edit friends and movies, but so far you can’t use them in one unified app. In this section, you’ll add a tab view to combine the interfaces. Tab views are a great way to let people quickly switch between a small number of related activities in your app. You’ll start by renaming ContentView to MovieList. Then you’ll make a new ContentView that contains the tabs.
