+++ FriendsFavoriteMovies/Views/ContentView.swift
@@ -2,75 +2,22 @@
 //  ContentView.swift
 //  FriendsFavoriteMovies
 //
-//  Created by Christian Manzaraz on 28/04/2024.
+//  Created by Christian Manzaraz on 30/04/2024.
 //
 
 import SwiftUI
-import SwiftData
 
 struct ContentView: View {
-    @Environment(\.modelContext) private var modelContext
-    @Query(sort: \Movie.title) private var movies: [Movie]
-    
-    @State private var newMovie: Movie?
-    
     var body: some View {
-        NavigationSplitView {
-            Group {
-                if !movies.isEmpty {
-                    List {
-                        ForEach(movies) { movie in
-                            NavigationLink {
-                                MovieDetail(movie: movie)
-                            } label: {
-                                Text(movie.title)
-                            }
-                        }
-                        .onDelete(perform: deleteItems)
-                    }
-                } else {
-                    ContentUnavailableView {
-                        Label("No Movies", systemImage: "film.stack")
-                    }
+        TabView {
+            MovieList()
+                .tabItem {
+                    Label("Movies", systemImage: "film.stack")
                 }
-            }
-            .navigationTitle("Movies")
-            .toolbar {
-                ToolbarItem(placement: .navigationBarTrailing) {
-                    EditButton()
+            FriendList()
+                .tabItem {
+                    Label("Friend", systemImage: "person.and.person")
                 }
-                ToolbarItem {
-                    Button(action: addMovie) {
-                        Label("Add Item", systemImage: "plus")
-                    }
-                }
-            }
-            .sheet(item: $newMovie) { movie in
-                NavigationStack {
-                    MovieDetail(movie: movie, isNew: true)
-                }
-                .interactiveDismissDisabled()
-            }
-        } detail: {
-            Text("Select a movie")
-                .navigationTitle("Movie")
-        }
-    }
-    
-    private func addMovie() {
-        withAnimation {
-            let newItem = Movie(title: "", releaseDate: .now)
-            modelContext.insert(newItem)
-            
-            newMovie = newItem
-        }
-    }
-    
-    private func deleteItems(offsets: IndexSet) {
-        withAnimation {
-            for index in offsets {
-                modelContext.delete(movies[index])
-            }
         }
     }
 }
@@ -78,9 +25,4 @@ struct ContentView: View {
 #Preview {
     ContentView()
         .modelContainer(SampleData.shared.modelContainer)
-}
-
-#Preview("Empty List") {
-    ContentView()
-        .modelContainer(for: Movie.self, inMemory: true)
 }
